Rename backend package into runner
Backend name is too generic and mostly related with the old runltp-ng
perl implementation. We want to create a specific package only for
runners, where a single runner is a class implementation which is used
to run commands on a specific target. For example, a target might be
a qemu instance, local host and remote host.