Refactor SerialRunner class
The stop method is not working properly because it's closing stdin and
stdout instead of stopping runner command. Also added ignore_echo into
the class constructor to ignore the first echo message when needed: this
is pretty common in Qemu instances, where prompt is sent back each time
we write into the stdin object.